# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/chats": {
      "post": {
        "operationId": "ChatsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChatDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatSummaryDto"
                }
              }
            }
          }
        },
        "summary": "Create a chat for the current principal",
        "tags": [
          "chats"
        ]
      },
      "get": {
        "operationId": "ChatsController_list",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "includeArchived",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatPageDto"
                }
              }
            }
          }
        },
        "summary": "List chats owned by the current principal",
        "tags": [
          "chats"
        ]
      }
    },
    "/api/v1/chats/{chatId}": {
      "get": {
        "operationId": "ChatsController_get",
        "parameters": [
          {
            "name": "chatId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatDetailDto"
                }
              }
            }
          }
        },
        "summary": "Get chat metadata and a page of messages",
        "tags": [
          "chats"
        ]
      },
      "patch": {
        "operationId": "ChatsController_update",
        "parameters": [
          {
            "name": "chatId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateChatDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatSummaryDto"
                }
              }
            }
          }
        },
        "summary": "Rename, archive, or change the future model for a chat",
        "tags": [
          "chats"
        ]
      },
      "delete": {
        "operationId": "ChatsController_delete",
        "parameters": [
          {
            "name": "chatId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "chats"
        ]
      }
    },
    "/api/v1/chats/{chatId}/messages/stream": {
      "post": {
        "operationId": "ChatsController_stream",
        "parameters": [
          {
            "name": "chatId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StreamMessageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Named SSE events ending in completion, cancellation, or error.",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string",
                  "example": "event: message.delta\ndata: {\"delta\":\"Hello\"}\n\n"
                }
              }
            }
          }
        },
        "summary": "Persist a user message and stream an Ollama response as SSE",
        "tags": [
          "chats"
        ]
      }
    },
...
  },
  "info": {
    "title": "Local LLM Chat API",
    "description": "HTTP API for the local-first LLM chat application.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access_token": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token"
      },
      "refresh_token": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refresh_token"
      }
    },
    "schemas": {
...
      "CreateChatDto": {
        "type": "object",
        "properties": {
          "model": {
            "type": "string",
            "example": "qwen2.5:1.5b",
            "maxLength": 200
          },
          "title": {
            "type": "string",
            "maxLength": 200
          },
          "firstPrompt": {
            "type": "string",
            "maxLength": 12000
          }
        },
        "required": [
          "model"
        ]
      },
      "ChatSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "selectedModel": {
            "type": "string"
          },
          "archivedAt": {
            "type": "object",
            "nullable": true
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "lastMessageAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "selectedModel",
          "createdAt",
          "updatedAt",
          "lastMessageAt"
        ]
      },
      "ChatPageDto": {
        "type": "object",
        "properties": {
          "chats": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChatSummaryDto"
            }
          },
          "nextCursor": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "chats"
        ]
      },
      "ChatMessageDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "assistant"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "streaming",
              "completed",
              "failed",
              "cancelled"
            ]
          },
          "content": {
            "type": "string"
          },
          "model": {
            "type": "object",
            "nullable": true
          },
          "tokenCount": {
            "type": "object",
            "nullable": true
          },
          "tokenCountSource": {
            "type": "string",
            "enum": [
              "ollama_reported",
              "estimated",
              "unknown"
            ]
          },
          "promptTokens": {
            "type": "object",
            "nullable": true
          },
          "completionTokens": {
            "type": "object",
            "nullable": true
          },
          "totalTokens": {
            "type": "object",
            "nullable": true
          },
          "finishReason": {
            "type": "object",
            "nullable": true
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "role",
          "status",
          "content",
          "tokenCountSource",
          "createdAt",
          "updatedAt"
        ]
      },
      "ChatDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "selectedModel": {
            "type": "string"
          },
          "archivedAt": {
            "type": "object",
            "nullable": true
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "lastMessageAt": {
            "type": "string"
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChatMessageDto"
            }
          },
          "nextCursor": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "id",
          "title",
          "selectedModel",
          "createdAt",
          "updatedAt",
          "lastMessageAt",
          "messages"
        ]
      },
      "UpdateChatDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200
          },
          "model": {
            "type": "string",
            "example": "qwen2.5:1.5b",
            "maxLength": 200
          },
          "archived": {
            "type": "boolean"
          }
        }
      },
      "StreamMessageDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "maxLength": 12000
          },
          "model": {
            "type": "string",
            "example": "qwen2.5:1.5b",
            "maxLength": 200
          }
        },
        "required": [
          "content"
        ]
      },
...
    }
  }
}